backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==3.2.2
certifi==2024.8.30
cffi==2.0.0
click==8.3.0
colorama==0.4.6
cryptography==46.0.1
deprecation==2.1.0
dnspython==2.8.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
fastapi==0.115.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
motor==3.5.1
packaging==25.0
passlib==1.7.4
postgrest==2.20.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.4.0
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.1
python-dotenv==1.0.1
python-jose==3.3.0
PyYAML==6.0.3
realtime==2.20.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.30.6
watchfiles==1.1.0
websockets==15.0.1
